requirement.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pypdf
requests
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu